Move files in `accounts` around to avoid nesting too deep
