+++ src/features/accounts/components/account/Account.js
@@ -1,11 +1,15 @@
-import PageTemplate from "../../../common/components/PageTemplate";
+import PageTemplate from "../../../../common/components/PageTemplate";
 import { Navigate, Outlet, useParams } from "react-router-dom";
-import Breadcrumbs from "../../../common/components/Breadcrumbs";
+import Breadcrumbs from "../../../../common/components/Breadcrumbs";
 import { Box, SimpleGrid } from "@chakra-ui/react";
-import { useReadAccountQuery } from "../../../app/api";
+import { useReadAccountQuery } from "../../../../app/api";
 import { format, formatDistanceToNow, parseISO } from "date-fns";
-import Card from "../../../common/components/Card";
+import Card from "../../../../common/components/Card";
 
+/**
+ * The page view for a specific account.
+ * Distinct from `AccountCard` which only displays a brief summary of the account.
+ */
 export default function Account() {
     const accountId = useParams().id;
     const { data, isLoading, isError } = useReadAccountQuery(accountId);
