+++ src/features/accounts/components/account/AccountDelete.js
@@ -1,7 +1,7 @@
 import { useNavigate, useOutletContext } from "react-router-dom";
 import { useToast } from "@chakra-ui/react";
-import FormModal from "../../../common/components/form/FormModal";
-import { useDeleteAccountMutation } from "../../../app/api";
+import FormModal from "../../../../common/components/form/FormModal";
+import { useDeleteAccountMutation } from "../../../../app/api";
 
 export default function AccountDelete() {
     const [deleteAccount] = useDeleteAccountMutation();
