+++ src/features/accounts/components/transactions/SpendingTransactionCreationForm.js
@@ -1,24 +1,27 @@
-import React from "react";
+import React, { useState } from "react";
 import * as Yup from "yup";
 import { Form, Formik } from "formik";
 import { VStack } from "@chakra-ui/react";
 import { useDispatch } from "react-redux";
-import { addTransaction } from "../state/transactions";
+import { addTransaction } from "../../state/transactions";
 import { format, formatISO, isMatch, parse } from "date-fns";
 import { useParams } from "react-router-dom";
 import FormTextField from "../../../../common/components/form/FormTextField";
+import TypeOfSpending from "./SpendingsDropDownField";
 
-export default function SpendingTransactionCreationForm({ afterSubmit }) {
+export default function ReceivingTransactionCreationForm({ afterSubmit }) {
     const { id } = useParams();
     const dispatch = useDispatch();
     const today = format(new Date(), "dd/MM/yyyy");
+    const [childData, setChildData] = useState("");
+
     return (
         <Formik
             initialValues={{
                 date: today,
-                amount: 123.9,
+                amount: -123.91,
                 balance: 4567.89,
-                description: "Some income.",
+                description: "An expense.",
             }}
             validationSchema={Yup.object({
                 date: Yup.string()
@@ -49,6 +52,7 @@ export default function SpendingTransactionCreationForm({ afterSubmit }) {
                         // Store dates as formatted ISO strings because Date objects aren't serializable.
                         date: formatISO(parse(date, "dd/MM/yyyy", new Date())),
                         accountId: parseInt(id),
+                        category: childData,
                     })
                 );
                 afterSubmit();
@@ -79,6 +83,11 @@ export default function SpendingTransactionCreationForm({ afterSubmit }) {
                         labelText="Balance"
                         placeholderText="4567.89"
                         helperText="This should be the balance of your account after the transaction."
+                    />
+                    <TypeOfSpending
+                        id="typeofspending"
+                        isRequired
+                        passChildData={setChildData}
                     />
                     <FormTextField
                         id="description"
