+++ src/features/accounts/components/upload/UploadModal.js
@@ -2,7 +2,7 @@ import UploadForm from "./UploadForm";
 import CancelButton from "../../../../common/components/buttons/CancelButton";
 import SubmitButton from "../../../../common/components/buttons/SubmitButton";
 import { useDispatch, useSelector } from "react-redux";
-import { deleteAllFiles, selectFiles } from "../state/files";
+import { deleteAllFiles, selectFiles } from "../../state/files";
 import { useNavigate } from "react-router-dom";
 import Modal from "../../../../common/components/Modal";
